Move relevant commands to MainViewModel
