+++ RedBookPlayer.GUI/ViewModels/PlayerViewModel.cs
@@ -236,19 +236,9 @@ namespace RedBookPlayer.GUI.ViewModels
         #region Commands
 
         /// <summary>
-        /// Command for handling keypresses
-        /// </summary>
-        public ReactiveCommand<KeyEventArgs, Unit> KeyPressCommand { get; }
-
-        /// <summary>
         /// Command for loading a disc
         /// </summary>
         public ReactiveCommand<Unit, Unit> LoadCommand { get; }
-
-        /// <summary>
-        /// Command for loading a disc from drag and drop
-        /// </summary>
-        public ReactiveCommand<DragEventArgs, Unit> LoadDragDropCommand { get; }
 
         #region Playback
 
@@ -355,9 +345,7 @@ namespace RedBookPlayer.GUI.ViewModels
         public PlayerViewModel()
         {
             // Initialize commands
-            KeyPressCommand = ReactiveCommand.Create<KeyEventArgs>(ExecuteKeyPress);
             LoadCommand = ReactiveCommand.Create(ExecuteLoad);
-            LoadDragDropCommand = ReactiveCommand.Create<DragEventArgs>(ExecuteLoadDragDrop);
 
             PlayCommand = ReactiveCommand.Create(ExecutePlay);
             PauseCommand = ReactiveCommand.Create(ExecutePause);
@@ -558,116 +546,6 @@ namespace RedBookPlayer.GUI.ViewModels
         }
 
         /// <summary>
-        /// Execute the result of a keypress
-        /// </summary>
-        public void ExecuteKeyPress(KeyEventArgs e)
-        {
-            // Open settings window
-            if(e.Key == App.Settings.OpenSettingsKey)
-            {
-                SettingsWindow settingsWindow = new SettingsWindow() { DataContext = App.Settings };
-                settingsWindow.Closed += OnSettingsClosed;
-                settingsWindow.ShowDialog(App.MainWindow);
-            }
-
-            // Load image
-            else if(e.Key == App.Settings.LoadImageKey)
-            {
-                ExecuteLoad();
-            }
-
-            // Toggle playback
-            else if(e.Key == App.Settings.TogglePlaybackKey || e.Key == Key.MediaPlayPause)
-            {
-                ExecuteTogglePlayPause();
-            }
-
-            // Stop playback
-            else if(e.Key == App.Settings.StopPlaybackKey || e.Key == Key.MediaStop)
-            {
-                ExecuteStop();
-            }
-
-            // Eject
-            else if(e.Key == App.Settings.EjectKey)
-            {
-                ExecuteEject();
-            }
-
-            // Next Track
-            else if(e.Key == App.Settings.NextTrackKey || e.Key == Key.MediaNextTrack)
-            {
-                ExecuteNextTrack();
-            }
-
-            // Previous Track
-            else if(e.Key == App.Settings.PreviousTrackKey || e.Key == Key.MediaPreviousTrack)
-            {
-                ExecutePreviousTrack();
-            }
-
-            // Next Index
-            else if(e.Key == App.Settings.NextIndexKey)
-            {
-                ExecuteNextIndex();
-            }
-
-            // Previous Index
-            else if(e.Key == App.Settings.PreviousIndexKey)
-            {
-                ExecutePreviousIndex();
-            }
-
-            // Fast Foward
-            else if(e.Key == App.Settings.FastForwardPlaybackKey)
-            {
-                ExecuteFastForward();
-            }
-
-            // Rewind
-            else if(e.Key == App.Settings.RewindPlaybackKey)
-            {
-                ExecuteRewind();
-            }
-
-            // Volume Up
-            else if(e.Key == App.Settings.VolumeUpKey || e.Key == Key.VolumeUp)
-            {
-                int increment = 1;
-                if(e.KeyModifiers.HasFlag(KeyModifiers.Control))
-                    increment *= 2;
-                if(e.KeyModifiers.HasFlag(KeyModifiers.Shift))
-                    increment *= 5;
-
-                ExecuteSetVolume(Volume + increment);
-            }
-
-            // Volume Down
-            else if(e.Key == App.Settings.VolumeDownKey || e.Key == Key.VolumeDown)
-            {
-                int decrement = 1;
-                if(e.KeyModifiers.HasFlag(KeyModifiers.Control))
-                    decrement *= 2;
-                if(e.KeyModifiers.HasFlag(KeyModifiers.Shift))
-                    decrement *= 5;
-
-                ExecuteSetVolume(Volume - decrement);
-            }
-
-            // Mute Toggle
-            else if(e.Key == App.Settings.ToggleMuteKey || e.Key == Key.VolumeMute)
-            {
-                ExecuteToggleMute();
-            }
-
-            // Emphasis Toggle
-            else if(e.Key == App.Settings.ToggleDeEmphasisKey)
-            {
-                ExecuteToggleDeEmphasis();
-            }
-        }
-
-        /// <summary>
         /// Load a disc image from a selection box
         /// </summary>
         public async void ExecuteLoad()
@@ -677,20 +555,6 @@ namespace RedBookPlayer.GUI.ViewModels
                 return;
 
             await LoadImage(path);
-        }
-
-        /// <summary>
-        /// Load the first valid drag-and-dropped disc image
-        /// </summary>
-        public async void ExecuteLoadDragDrop(DragEventArgs e)
-        {
-            IEnumerable<string> fileNames = e.Data.GetFileNames();
-            foreach(string filename in fileNames)
-            {
-                bool loaded = await LoadImage(filename);
-                if(loaded)
-                    break;
-            }
         }
 
         /// <summary>
@@ -922,11 +786,6 @@ namespace RedBookPlayer.GUI.ViewModels
             App.PlayerView.DataContext = this;
             App.PlayerView.ViewModel = this;
         }
-
-        /// <summary>
-        /// Handle the settings window closing
-        /// </summary>
-        private void OnSettingsClosed(object sender, EventArgs e) => RefreshFromSettings();
 
         /// <summary>
         /// Update the view-model from the Player
